% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countCompoundTypes.R
\name{categoryCuts}
\alias{categoryCuts}
\title{Count the number of occurances inside a range}
\usage{
categoryCuts(xx, ranges = list(OtoC = c(0, 1)), inclusive = list(OtoC =
  c(TRUE, TRUE)), na.rm = TRUE)
}
\arguments{
\item{xx}{a dataframe with names matching those in ranges and inclusive. Values within range to be counted}

\item{ranges}{a list of arrays of length 2 with the (min, max) cut off values}

\item{inclusive}{a list of arrays of length 2 with the flags to include (lower, upper) cut off values in rante}

\item{na.rm}{ignore na values}
}
\value{
a count times the value in xx fell within range of interest
}
\description{
Count the number of occurances inside a range
}

