% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeVanKrevlen.R
\name{makeVanKrevelen}
\alias{makeVanKrevelen}
\title{Create a Van Krevlen plot}
\usage{
makeVanKrevelen(fileIn = "FTICRInputFile.csv", massHeader = c("Mass",
  "m.z"), ratioHeaders = c("C", "H", "O", "N", "X13C", "S", "P", "C13"),
  sampleRegStr = "(X.out)|(^X\\\\d+$)|(std)|(IntCal_)", samplesToPlot = 1:9,
  colorBy = c("SN", "intensity")[2], verbose = TRUE)
}
\arguments{
\item{fileIn}{csv file with the FT-ICR-MS data}

\item{massHeader}{array of strings with the names of the mass to charge ratio column}

\item{ratioHeaders}{array of strings with the names of the elemental count columns}

\item{sampleRegStr}{regular expression defining the sample columns}

\item{samplesToPlot}{number of samples to plot}

\item{colorBy}{string to color plot by intensity or by inclusion of S or N in element assignment}

\item{verbose}{boolean flag for verbose outputs}
}
\value{
a ggplot object with the graph
}
\description{
Create a Van Krevlen plot
}

