% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalizeMassNotation.R
\name{normalizeMassNotation}
\alias{normalizeMassNotation}
\title{Standardize mass notation}
\usage{
normalizeMassNotation(data.df, sampleIDs, elementKey = list(C = "C", H = "H",
  O = "O", N = "N", S = "S", P = "P"), verbose = FALSE)
}
\arguments{
\item{data.df}{data frame with samples and elements}

\item{sampleIDs}{an array of strings with the sample column names}

\item{elementKey}{a list with names that match C, H, O, N, S, P with the associated column names. Note that if you want to count C+C13 then set list(C=c('C', 'C13')).}

\item{verbose}{a boolean to flag lots of output that is hopefully informative}
}
\value{
data.frame with indecies and ratios of interest plus the standardize element names and sampleIDs
}
\description{
Force C, H, O, N, S, and P for the headers and calculate various ratios and indecies of interest.
}

