% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFTICR.R
\name{readFTICR}
\alias{readFTICR}
\title{Read in chuncks of an FT-ICR-MS file}
\usage{
readFTICR(fileIn, massHeader = c("Mass", "m.z"),
  sampleRegStr = "(X.out)|(^X\\\\d+$)|(std)|(IntCal_)", samplesToRead,
  verbose = FALSE)
}
\arguments{
\item{fileIn}{a string identifying the csv file to read our FT-ICR-MS data from}

\item{massHeader}{an array of strings or single string naming the mass to charge header}

\item{sampleRegStr}{a regular expression matching the sample headers (and ONLY the sample headers)}

\item{samplesToRead}{an array of indecies or regular expression or array of names of the samples to be read in}

\item{verbose}{boolean flag for verbose outputs}
}
\value{
a data frame with the long table format with the  mass ID, sample ID, and intensity
}
\description{
Read in chuncks of an FT-ICR-MS file
}
\examples{
##read samples by index (note this is the Nth sample not the Nth column)
#readFTICR(fileIn='testfile.csv', samplesToRead=1:2)
##read samples by regular expression match of sample name
#readFTICR(fileIn='testFile.csv', samplesToRead='out2\\\\d')
##read samples by matched name
#readFTICR(fileIn='testFile.csv', samplesToRead=c('X.out20', 'X.out21'))
}

