% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readMass.R
\name{readMass}
\alias{readMass}
\title{Read in characteristics of the mass to charge ratios}
\usage{
readMass(fileIn, massHeader = "Mass", elementKey = list(C = "C", H = "H", O
  = "O", N = "N", S = "S", P = "P"), verbose = FALSE)
}
\arguments{
\item{fileIn}{a string identifying the csv file to read our data from}

\item{massHeader}{an array of strings or single string naming the mass to charge header}

\item{elementKey}{a list of all the column names to be counted for each element}

\item{verbose}{boolean flag for verbose outputs}
}
\value{
a data frame with the mass, orginal elements [if names different from standard],
element count, and metrics [if appropreate elements are present]
}
\description{
Read in characteristics of the mass to charge ratios
}
\examples{
##massHeader can have extra entries that will be ignored
#readMass(fileIn='testfile.csv', massHeader=c('m.z', 'mass', 'nonsence'))

##elementKey can contain isotopes which will be included in the final element count
#readMass(fileIn='testfile.csv', massHeader='m.z', elementKey=list(C=c('C13', 'C'),
#H='H', O='O', N=c('N', 'N15'), S='S', P='P'))
}

